diff: update GH comment when diff public URL changes
Until now we considered a Diff comment to be digested by its summary
only, however this is confusing for users because they don't see later
updates when there's no structural changes between a first diff
comment and a later commit.

With this commit we include the diff public url to be part of the
digest to make sure we update the bot GH comment when the diff
changes.

Thank you @gmourier for mentioning this issue during your tests in
https://github.com/meilisearch/specifications/pull/58 ! 😊