diff: use pull request base ref to perform Bump Diff
_Requirements: Bump CLI needs to be updated to v2.2.2 to support two
files input. cf https://github.com/bump-sh/cli/pull/141_

The current 'diff' command of this action will only be able to produce
a diff with the currently deployed API definition file on
bump.sh. This is nice when you merge API changes in a single branch
one after another. However in the context of multiple parallel
changes, and with the notion of “staging” API definition file (file not yet
deployed but which is considered as the most up-to-date base), this
diff process doesn't make much sense. Why? because it will include all
changes made to the “staging” version in your own PR change.

Thus, this PR reads the PR base ref to be able to restore (from git)
the base (target) branch API definition file version. With this target
branch file version we can perform a “real” diff between the current
version file and the target branch file.