diff: instead of using target base commit, use last common ancestor
We've been using the base branch version to be able compute the PR
diff, but this is wrong! We need to check the diff between the last
common ancestor and the HEAD of the branch.

With this we will be able to have a correct diff to compute only the
real PR's files diff.