diff: make sure to delete comments when no version was created
This is a follow-up to #165 to make sure to delete an existing Github
bot comment if there is no diff.

The previous #165 only handled the cases where a non structural change
exists, the changes present in this commit is for completely empty
diffs (I.e. no changes with the currently deployed version).