diff: don't run with inexistant restored file
When preparing the two files to compare with the Diff action, we need
to check if the restored file exist (in case the current PR is adding
the file, the previous base file will not exist).

If the base file doesn't exist, then don't try to add it in the
diff command paramters.