fix: make sure to restore all repository for diff command
API contract files can be split out in multiple files to favor
reusability and readability. This is done thanks to the `$ref`
[references objects](https://spec.openapis.org/oas/v3.1.0#reference-object).

For our `diff` command we need to restore the previous version of the
contract file to be able to compare previous with current HEAD. The
current implementation would only restore the given root file.
With this PR we restore the entire repository in its merge-base
version. This will avoid any errors while needing to compare two API
contract files which are written with external reference objects.

Fixes #179