fix(diff): add “doc” related context in GH diff comments
Before this commit all comments published to GH on the diff feature
would be independent from the potential documentation/hub id (or slug)
given to the GH action inputs.

However, some of our users may want to use the GH action on multiple
different specification files (attached to different docs).

Thus with this commit we add the “context” of the documentation in the
comment signature. In this way we can remove only previous comments
for the given documentation (and avoid deleting any of the other Bump
comments)