inputs: make the `doc` parameter optional
The GH action `doc:` input parameter should be optional.

Indeed, since we introduced the [directory deploy for Hubs in the CLI](https://github.com/bump-sh/cli/pull/462), the `--doc`
parameter became optional for hubs deployments.

This PR makes sure the `doc` input of the GitHub action is also
optional to be able to deploy a whole directory to a Bump.sh Hub from
the GitHub action too.