diff: make sure to differentiate No Change vs Content change only
This commit makes sure to have a different behavior during a diff
between a No Change at all result, vs a No Content Change result.

Before this commit we would only display a “No diff found” message in
the GH action logs.

With this new commit, if there's a content change detected we now
publish a comment containing a link to be able to preview the change.