fix(deploy): no need for quoting around `filename_pattern` input val
Not sur why I added some quotes around the newly introduced
`filename_pattern` input, as we don't quote any of the other input
values.

This led to a bug of double single quotes, thus the default pattern
for the GH action to be the litteral `'{slug}-api'` (with the single
quotes).

Sorry for that!