fix: use doc context in github comment docDigest
This commit includes more information in the current “docDigest” which
is used to differentiate github diff comments for each docs that are
diffed with the github action.

We had the “hub” context encoded in the digest, but forgot the
“branch” name. This could lead to an inconsistant comment if one tries
to execute two diff commands, with two differnt branch name. Only the
second diff comment would be visible on the PR (instead of two
comments one for each branch).

The commit also moves the “doc digest” computation inside the `Repo`
class instead of the main code.